Single-path Pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp; kwargs...)
pathfinder(logp, ∇logp; kwargs...)
pathfinder(fun::GalacticOptim::OptimizationFunction; kwargs...)
pathfinder(prob::GalacticOptim::OptimizationProblem; kwargs...)

Find the best multivariate normal approximation encountered while maximizing logp.

From an optimization trajectory, Pathfinder constructs a sequence of (multivariate normal) approximations to the distribution specified by logp. The approximation that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between the approximation and the true distribution, is returned.

The covariance of the multivariate normal distribution is an inverse Hessian approximation constructed using at most the previous history_length steps.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp. If not provided, logp is automatically differentiated using the backend specified in ad_backend.

fun::GalacticOptim.OptimizationFunction: an optimization function that represents -logp(x) with its gradient. It must have the necessary features (e.g. a Hessian function) for the chosen optimization algorithm. For details, see GalacticOptim.jl: OptimizationFunction.

prob::GalacticOptim.OptimizationProblem: an optimization problem containing a function with the same properties as fun, as well as an initial point, in which case init and dim are ignored.

Keywords

dim::Int: dimension of the target distribution. If not provided, init or must be. Ignored if init is provided.

init::AbstractVector{<:Real}: initial point of length dim from which to begin optimization. If not provided, an initial point of type Vector{Float64} and length dim is created and filled using init_sampler.

init_scale::Real: scale factor $s$ such that the default init_sampler samples entries uniformly in the range $[-s, s]$

init_sampler: function with the signature (rng, x) -> x that modifies a vector of length dims in-place to generate an initial point

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

ndraws::Int=ndraws_elbo: number of approximate draws to return

ad_backend=AD.ForwardDiffBackend(): AbstractDifferentiation.jl AD backend.

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

executor::Transducers.Executor=Transducers.SequentialEx(): Transducers.jl executor that determines if and how to perform ELBO computation in parallel. The default (SequentialEx()) performs no parallelization. If rng is known to be thread-safe, and the log-density function is known to have no internal state, then Transducers.PreferParallel() may be used to parallelize log-density evaluation. This is generally only faster for expensive log density functions.

history_length::Int=6: Size of the history used to approximate the inverse Hessian.

optimizer: Optimizer to be used for constructing trajectory. Can be any optimizer compatible with GalacticOptim, so long as it supports callbacks. Defaults to Optim.LBFGS(; m=history_length, linesearch=LineSearches.MoreThuente()). See the GalacticOptim.jl documentation for details.

ntries::Int=1_000: Number of times to try the optimization, restarting if it fails. Before every restart, a new initial point is drawn using init_sampler.

kwargs... : Remaining keywords are forwarded to GalacticOptim.solve.

Returns

PathfinderResult

Pathfinder.PathfinderResult — Type

PathfinderResult

Container for results of single-path Pathfinder.

Fields

input: User-provided input object, e.g. either logp, (logp, ∇logp), optim_fun, optim_prob, or another object.

optimizer: Optimizer used for maximizing the log-density

rng: Pseudorandom number generator that was used for sampling

optim_prob::GalacticOptim.OptimizationProblem: Otimization problem used for optimization

logp: Log-density function

fit_distribution::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

draws::AbstractMatrix{<:Real}: draws from multivariate normal with size (dim, ndraws)

fit_distribution_transformed: fit_distribution transformed to the same space as the user-supplied target distribution. This is only different from fit_distribution when integrating with other packages, and its type depends on the type of input.

draws_transformed: draws transformed to be draws from fit_distribution_transformed.

fit_iteration::Int: Iteration at which ELBO estimate was maximized

num_tries::Int: Number of tries until Pathfinder succeeded

optim_solution::GalacticOptim.OptimizationSolution: Solution object of optimization.

optim_trace::Pathfinder.OptimizationTrace: container for optimization trace of points, log-density, and gradient. The first point is the initial point.

fit_distributions::AbstractVector{Distributions.MvNormal}: Multivariate normal distributions for each point in optim_trace, where fit_distributions[fit_iteration + 1] == fit_distribution

elbo_estimates::AbstractVector{<:Pathfinder.ELBOEstimate}: ELBO estimates for all but the first distribution in fit_distributions.

Returns

PathfinderResult

Examples

For a simple example, we'll run Pathfinder on a multivariate normal distribution with a dense covariance matrix.

using LinearAlgebra, Pathfinder, Random
Random.seed!(42)

Σ = [
    2.71   0.5    0.19   0.07   1.04
    0.5    1.11  -0.08  -0.17  -0.08
    0.19  -0.08   0.26   0.07  -0.7
    0.07  -0.17   0.07   0.11  -0.21
    1.04  -0.08  -0.7   -0.21   8.65
]
μ = [-0.55, 0.49, -0.76, 0.25, 0.94]
P = inv(Symmetric(Σ))

function logp(x)
    z = x - μ
    return -dot(z, P, z) / 2;
end
∇logp(x) = P * (μ - x)

Now we run Pathfinder.

julia> result = pathfinder(logp, ∇logp; dim=5, ndraws=100, ndraws_elbo=100)Single-path Pathfinder result
  tries: 1
  draws: 100
  fit iteration: 34 (total: 38)
  fit ELBO: 3.79 ± 0.04
  fit distribution: Distributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}(
dim: 5
μ: [-0.5499999949158282, 0.4899999988178169, -0.7599999979862843, 0.2499999989895235, 0.9400000067909953]
Σ: [2.8053766005274547 0.46197980209657374 … 0.04942703113879024 1.204024834879955; 0.46197980209657374 1.1203155589697753 … -0.1502115458530225 -0.20284056239515952; … ; 0.04942703113879024 -0.1502115458530225 … 0.12093759888809286 -0.30434396827746313; 1.204024834879955 -0.20284056239515952 … -0.30434396827746313 9.369064525618189]
)

result is a PathfinderResult. See its docstring for a description of its fields.

result.fit_distribution is a multivariate normal approximation to our target distribution. Its mean and covariance are quite close to our target distribution's.

julia> result.fit_distribution.μ5-element Vector{Float64}:
 -0.5499999949158282
  0.4899999988178169
 -0.7599999979862843
  0.2499999989895235
  0.9400000067909953
julia> result.fit_distribution.Σ5×5 Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}:
 2.80538    0.46198     0.21003     0.049427    1.20402
 0.46198    1.12032    -0.0846231  -0.150212   -0.202841
 0.21003   -0.0846231   0.279982    0.0831588  -0.769305
 0.049427  -0.150212    0.0831588   0.120938   -0.304344
 1.20402   -0.202841   -0.769305   -0.304344    9.36906

result.draws is a Matrix whose columns are the requested draws from result.fit_distribution:

julia> result.draws5×100 Matrix{Float64}:
  1.03914   -0.235486   2.35216   …  -1.72906   -2.68335   -1.50072
 -0.305742  -1.14926   -0.524424     -1.09469   -1.75937    3.06293
 -1.13874   -0.41053    0.563586      0.252442  -0.169261  -0.572252
  0.704461   0.765028   0.554606      1.61316    0.562983  -0.00690929
  7.4157    -3.51686   -1.99529      -3.55038    0.660964  -0.465471