Pathfinder

Documentation for Pathfinder.

Pathfinder.WoodburyPDMat

Pathfinder.multipathfinder

Pathfinder.pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp, ∇logp, θ₀::AbstractVector{<:Real}, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while optimizing logp.

The multivariate normal approximation returned is the one that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀: initial point from which to begin optimization

ndraws: number of approximate draws to return

Keywords

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

optimizer::Optim.AbstractOptimizer: Optimizer to be used for constructing trajectory.

Defaults to Optim.LBFGS(; m=5, linesearch=LineSearches.MoreThuente()).

history_length::Int=5: Size of the history used to approximate the inverse Hessian.

This should only be set when optimizer is not an Optim.LBFGS.

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

kwargs... : Remaining keywords are forwarded to Optim.Options.

Returns

q::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws draws from multivariate normal

logqϕ::Vector{<:Real}: log-density of multivariate normal at ϕ values

Pathfinder.multipathfinder — Function

multipathfinder(
    logp,
    ∇logp,
    θ₀s::AbstractVector{AbstractVector{<:Real}},
    ndraws::Int;
    kwargs...
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

For n=length(θ₀s), n parallel runs of pathfinder produce n multivariate normal approximations of the posterior. These are combined to a mixture model with uniform weights.

Draws from the components are then resampled with replacement. If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀s: vector of initial points from which each optimization will begin

ndraws: number of approximate draws to return

Keywords

ndraws_per_run::Int=5: The number of draws to take for each component before resampling.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws approxiate draws from target distribution

Pathfinder.WoodburyPDMat — Type

WoodburyPDMat <: PDMats.AbstractPDMat

Lazily represents a real positive definite (PD) matrix as an update to a full-rank PD matrix.

WoodburyPDMat(A, B, D)

Constructs the $n \times n$ PD matrix

\[W = A + B D B^\mathrm{T},\]

where $A$ is an $n \times n$ full rank positive definite matrix, $D$ is an $m \times m$ symmetric matrix, and $B$ is an $n \times m$ matrix. Note that $B$ and $D$ must be chosen such that $W$ is positive definite; this is only implicitly checked.

Overloads for WoodburyPDMat make extensive use of the following decomposition. Let $L_A L_A^\mathrm{T} = A$ be the Cholesky decomposition of $A$, and let $Q R = L_A^{-1} B$ be a thin QR decomposition. Define $C = I + RDR^\mathrm{T}$, with the Cholesky decomposition $L_C L_C^\mathrm{T} = C$. Then, $W = T T^\mathrm{T}$, where

\[T = L_A Q \begin{pmatrix} L_C & 0 \\ 0 & I \end{pmatrix}.\]

The positive definite requirement is equivalent to the requirement that both $A$ and $C$ are positive definite.

For a derivation of this decomposition for the special case of diagonal $A$, see appendix A of [Zhang2021].

Zhang2021Lu Zhang, Bob Carpenter, Andrew Gelman, Aki Vehtari (2021). Pathfinder: Parallel quasi-Newton variational inference. arXiv: 2108.03782 [stat.ML]