Multi-path Pathfinder

Pathfinder.multipathfinder — Function

multipathfinder(
    logp,
    [∇logp,]
    θ₀s::AbstractVector{AbstractVector{<:Real}},
    ndraws::Int;
    kwargs...
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

For nruns=length(θ₀s), nruns parallel runs of pathfinder produce nruns multivariate normal approximations $q_k = q(\phi | \mu_k, \Sigma_k)$ of the posterior. These are combined to a mixture model $q$ with uniform weights.

$q$ is augmented with the component index to generate random samples, that is, elements $(k, \phi)$ are drawn from the augmented mixture model

\[\tilde{q}(\phi, k | \mu, \Sigma) = K^{-1} q(\phi | \mu_k, \Sigma_k),\]

where $k$ is a component index, and $K=$ nruns. These draws are then resampled with replacement. Discarding $k$ from the samples would reproduce draws from $q$.

If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution $p$ instead of $q$.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp. If not provided, logp is automatically differentiated using the backend specified in ad_backend.

θ₀s::AbstractVector{AbstractVector{<:Real}}: vector of length nruns of initial points of length dim from which each single-path Pathfinder run will begin

ndraws::Int: number of approximate draws to return

Keywords

ad_backend=AD.ForwardDiffBackend(): AbstractDifferentiation.jl AD backend.

ndraws_per_run::Int=5: The number of draws to take for each component before resampling.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

kwargs... : Remaining keywords are forwarded to pathfinder.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::AbstractMatrix{<:Real}: approximate draws from target distribution with size (dim, ndraws)

component_inds::Vector{Int}: Indices $k$ of components in $q$ from which each column in ϕ was drawn.

multipathfinder(
    f::GalacticOptim.OptimizationFunction,
    θ₀s::AbstractVector{<:Real},
    ndraws::Int;
    kwargs...,
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

f is a user-created optimization function that represents the negative log density with its gradient and must have the necessary features (e.g. a Hessian function or specified automatic differentiation type) for the chosen optimization algorithm. For details, see GalacticOptim.jl: OptimizationFunction.

See multipathfinder for a description of remaining arguments.

Examples

Especially for complicated target distributions, it's more useful to run multi-path Pathfinder. One difficult distribution to sample is Neal's funnel:

\[\tau \sim \mathrm{Normal}(0, 3)\\ \beta_i \sim \mathrm{Normal}(0, e^{\tau/2})\]

Such funnel geometries appear in other models and typically frustrate MCMC sampling. Multi-path Pathfinder can't sample the funnel well, but it can quickly give us draws that can help us diagnose that we have a funnel.

In this example, we draw from a 100-dimensional funnel and visualize 2 dimensions.

using ForwardDiff, Pathfinder, Random

Random.seed!(42)

function logp(x)
    n = length(x)
    τ = x[1]
    β = view(x, 2:n)
    return ((τ / 3)^2 + (n - 1) * τ + sum(b -> abs2(b * exp(-τ / 2)), β)) / -2
end
∇logp(x) = ForwardDiff.gradient(logp, x)

dim = 100
ndraws = 1_000
nruns = 20

Next, we create initial points for the Pathfinder runs.

julia> θ₀s = collect.(eachcol(rand(dim, nruns) .* 20 .- 10));

Now we run multi-path Pathfinder.

julia> ndraws_per_run = ndraws ÷ nruns50
julia> @time q, ϕ, component_ids = multipathfinder(logp, ∇logp, θ₀s, ndraws; ndraws_per_run);[ Info: Optimized for 906 iterations. Maximum ELBO of 82.5 reached at iteration 11.
[ Info: Optimized for 828 iterations. Maximum ELBO of 85.8 reached at iteration 7.
[ Info: Optimized for 837 iterations. Maximum ELBO of 86.85 reached at iteration 8.
[ Info: Optimized for 865 iterations. Maximum ELBO of 84.88 reached at iteration 11.
[ Info: Optimized for 28 iterations. Maximum ELBO of 59.19 reached at iteration 8.
[ Info: Optimized for 844 iterations. Maximum ELBO of 87.56 reached at iteration 10.
[ Info: Optimized for 827 iterations. Maximum ELBO of 87.49 reached at iteration 8.
[ Info: Optimized for 806 iterations. Maximum ELBO of 72.48 reached at iteration 3.
[ Info: Optimized for 841 iterations. Maximum ELBO of 81.7 reached at iteration 9.
[ Info: Optimized for 865 iterations. Maximum ELBO of 86.16 reached at iteration 9.
[ Info: Optimized for 958 iterations. Maximum ELBO of 78.36 reached at iteration 9.
[ Info: Optimized for 28 iterations. Maximum ELBO of 82.77 reached at iteration 16.
[ Info: Optimized for 835 iterations. Maximum ELBO of 83.91 reached at iteration 13.
[ Info: Optimized for 897 iterations. Maximum ELBO of 53.71 reached at iteration 11.
[ Info: Optimized for 865 iterations. Maximum ELBO of 88.19 reached at iteration 12.
[ Info: Optimized for 827 iterations. Maximum ELBO of -58.59 reached at iteration 8.
[ Info: Optimized for 834 iterations. Maximum ELBO of 86.41 reached at iteration 12.
[ Info: Optimized for 867 iterations. Maximum ELBO of 55.08 reached at iteration 13.
[ Info: Optimized for 807 iterations. Maximum ELBO of 83.34 reached at iteration 4.
[ Info: Optimized for 29 iterations. Maximum ELBO of 79.14 reached at iteration 20.
┌ Warning: Pareto shape k = 0.89 > 0.7. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/ZPzVl/src/core.jl:264
 25.860422 seconds (52.91 M allocations: 4.516 GiB, 7.20% gc time, 82.06% compilation time)

The first return value is a uniformly-weighted Distributions.MixtureModel. Each component is the result of a single Pathfinder run.

julia> typeof(q)Distributions.MixtureModel{Distributions.Multivariate, Distributions.Continuous, Distributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}, Distributions.Categorical{Float64, Vector{Float64}}}

While we could draw samples from q directly, these aren't equivalent to the samples returned by multi-path Pathfinder, which uses multiple importance sampling with Pareto-smoothed importance resampling to combine the individual runs.

Note that PSIS.jl, which smooths the importance weights, warns us that the importance weights are unsuitable for computing estimates, so we should definitely run MCMC to get better samples. Pathfinder's samples can still be useful though for initializing MCMC.

Let's compare Pathfinder's samples with samples from q directly, plotted over the exact marginal density.

using Plots

τ_approx = ϕ[1, :]
β₁_approx = ϕ[2, :]
ϕ2 = rand(q, ndraws)
τ_approx2 = ϕ2[1, :]
β₁_approx2 = ϕ2[2, :]

τ_range = -15:0.01:5
β₁_range = -5:0.01:5

plt = contour(β₁_range, τ_range, (β, τ) -> exp(logp([τ, β])))
scatter!(β₁_approx2, τ_approx2; msw=0, ms=2, alpha=0.3, color=:blue, label="Pathfinder w/o IR")
scatter!(β₁_approx, τ_approx; msw=0, ms=2, alpha=0.5, color=:red, label="Pathfinder")
plot!(xlims=extrema(β₁_range), ylims=extrema(τ_range), xlabel="β₁", ylabel="τ", legend=true)
plt

As expected from PSIS.jl's warning, neither set of samples cover the density well. But we can also see that the mixture model (q) places too many samples in regions of low density, which is corrected by the importance resampling.

We can check how much each component contributed to the returned sample.

histogram(component_ids; bins=(0:nruns) .+ 0.5, bar_width=0.8, xticks=1:nruns,
          xlabel="Component index", ylabel="Count", legend=false)