Multi-path Pathfinder

Pathfinder.multipathfinder — Function

multipathfinder(
    logp,
    ∇logp,
    θ₀s::AbstractVector{AbstractVector{<:Real}},
    ndraws::Int;
    kwargs...
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

For nruns=length(θ₀s), nruns parallel runs of pathfinder produce nruns multivariate normal approximations $q_k = q(\phi | \mu_k, \Sigma_k)$ of the posterior. These are combined to a mixture model $q$ with uniform weights.

$q$ is augmented with the component index to generate random samples, that is, elements $(k, \phi)$ are drawn from the augmented mixture model

\[\tilde{q}(\phi, k | \mu, \Sigma) = K^{-1} q(\phi | \mu_k, \Sigma_k),\]

where $k$ is a component index, and $K=$ nruns. These draws are then resampled with replacement. Discarding $k$ from the samples would reproduce draws from $q$.

If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution $p$ instead of $q$.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀s: vector of length nruns of initial points of length dim from which each single-path Pathfinder run will begin

ndraws: number of approximate draws to return

Keywords

ndraws_per_run::Int=5: The number of draws to take for each component before resampling.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::AbstractMatrix{<:Real}: approximate draws from target distribution with size (dim, ndraws)

component_inds::Vector{Int}: Indices $k$ of components in $q$ from which each column

in ϕ was drawn.

Examples

Especially for complicated target distributions, it's more useful to run multi-path Pathfinder. One difficult distribution to sample is Neal's funnel:

\[\tau \sim \mathrm{Normal}(0, 3)\\ \beta_i \sim \mathrm{Normal}(0, e^{\tau/2})\]

Such funnel geometries appear in other models and typically frustrate MCMC sampling. Multi-path Pathfinder can't sample the funnel well, but it can quickly give us draws that can help us diagnose that we have a funnel.

In this example, we draw from a 100-dimensional funnel and visualize 2 dimensions.

using ForwardDiff, Pathfinder, Random

Random.seed!(42)

function logp(x)
    n = length(x)
    τ = x[1]
    β = view(x, 2:n)
    return ((τ / 3)^2 + (n - 1) * τ + sum(b -> abs2(b * exp(-τ / 2)), β)) / -2
end
∇logp(x) = ForwardDiff.gradient(logp, x)

dim = 100
ndraws = 1_000
nruns = 20

Next, we create initial points for the Pathfinder runs.

julia> θ₀s = collect.(eachcol(rand(dim, nruns) .* 20 .- 10));

Now we run multi-path Pathfinder.

julia> ndraws_per_run = ndraws ÷ nruns50
julia> @time q, ϕ, component_ids = multipathfinder(logp, ∇logp, θ₀s, ndraws; ndraws_per_run);[ Info: Optimized for 934 iterations. Maximum ELBO of 89.21 reached at iteration 10.
[ Info: Optimized for 825 iterations. Maximum ELBO of 63.9 reached at iteration 18.
[ Info: Optimized for 845 iterations. Maximum ELBO of 44.1 reached at iteration 12.
[ Info: Optimized for 899 iterations. Maximum ELBO of 53.27 reached at iteration 7.
[ Info: Optimized for 830 iterations. Maximum ELBO of 86.41 reached at iteration 6.
[ Info: Optimized for 853 iterations. Maximum ELBO of 74.37 reached at iteration 10.
[ Info: Optimized for 26 iterations. Maximum ELBO of 85.55 reached at iteration 5.
[ Info: Optimized for 874 iterations. Maximum ELBO of 84.47 reached at iteration 13.
[ Info: Optimized for 832 iterations. Maximum ELBO of 86.64 reached at iteration 10.
[ Info: Optimized for 825 iterations. Maximum ELBO of 77.89 reached at iteration 10.
[ Info: Optimized for 380 iterations. Maximum ELBO of 86.97 reached at iteration 5.
[ Info: Optimized for 853 iterations. Maximum ELBO of 86.88 reached at iteration 4.
[ Info: Optimized for 838 iterations. Maximum ELBO of 78.08 reached at iteration 10.
[ Info: Optimized for 833 iterations. Maximum ELBO of 83.8 reached at iteration 9.
[ Info: Optimized for 830 iterations. Maximum ELBO of 43.72 reached at iteration 8.
[ Info: Optimized for 856 iterations. Maximum ELBO of 70.19 reached at iteration 16.
[ Info: Optimized for 848 iterations. Maximum ELBO of 79.95 reached at iteration 16.
[ Info: Optimized for 837 iterations. Maximum ELBO of 81.6 reached at iteration 12.
[ Info: Optimized for 437 iterations. Maximum ELBO of 61.3 reached at iteration 12.
[ Info: Optimized for 844 iterations. Maximum ELBO of 88.62 reached at iteration 12.
┌ Warning: Pareto k statistic exceeded 0.7. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/KhMUf/src/PSIS.jl:100
 27.684529 seconds (42.71 M allocations: 3.487 GiB, 6.93% gc time, 0.02% compilation time)

The first return value is a uniformly-weighted Distributions.MixtureModel. Each component is the result of a single Pathfinder run.

julia> typeof(q)Distributions.MixtureModel{Distributions.Multivariate, Distributions.Continuous, Distributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}, Distributions.Categorical{Float64, Vector{Float64}}}

Note that while we could draw samples from q directly, these aren't equivalent to the samples returned by multi-path Pathfinder, which uses multiple importance sampling with Pareto-smoothed importance resampling to combine the individual runs. Let's draw exact samples from the funnel and compare them with Pathfinder's samples and samples from q directly.

using Plots

τ = randn(ndraws) * 3
β₁ = @. randn() * exp(τ / 2)
τ_approx = ϕ[1, :]
β₁_approx = ϕ[2, :]
ϕ2 = rand(q, ndraws)
τ_approx2 = ϕ2[1, :]
β₁_approx2 = ϕ2[2, :]

plt = scatter(β₁, τ; msw=0, ms=2, alpha=0.1, label="Exact")
scatter!(β₁_approx2, τ_approx2; msw=0, ms=2, alpha=0.2, label="multi-path Pathfinder w/o IR")
scatter!(β₁_approx, τ_approx; msw=0, ms=2, alpha=0.2, label="multi-path Pathfinder")
plot!(xlims=(-15, 15), ylims=(-15, 15), xlabel="β₁", ylabel="τ", legend=true)
plt

Here we can see that the mixture model (q) places too much probability mass on the lower part of the funnel, which is corrected by the importance resampling.

We can check how much each component contributed to the returned sample.

histogram(component_ids; bins=(0:nruns) .+ 0.5, bar_width=0.8, xticks=1:nruns,
          xlabel="Component index", ylabel="Count", legend=false)