Multi-path Pathfinder

Pathfinder.multipathfinder — Function

multipathfinder(
    logp,
    ∇logp,
    θ₀s::AbstractVector{AbstractVector{<:Real}},
    ndraws::Int;
    kwargs...
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

For n=length(θ₀s), n parallel runs of pathfinder produce n multivariate normal approximations of the posterior. These are combined to a mixture model with uniform weights.

Draws from the components are then resampled with replacement. If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀s: vector of initial points from which each optimization will begin

ndraws: number of approximate draws to return

Keywords

ndraws_per_run::Int=5: The number of draws to take for each component before resampling.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws approxiate draws from target distribution

Examples

Especially for complicated target distributions, it's more useful to run multi-path Pathfinder. One difficult distribution to sample is Neal's funnel:

\[\tau \sim \mathrm{Normal}(0, 3)\\ \beta_i \sim \mathrm{Normal}(0, e^{\tau/2})\]

Such funnel geometries appear in other models and typically frustrate MCMC sampling. Multi-path Pathfinder can't sample the funnel well, but it can quickly give us draws that can help us diagnose that we have a funnel.

In this example, we draw from a 100-dimensional funnel and visualize 2 dimensions.

using ForwardDiff, Pathfinder, Random

Random.seed!(42)

function logp(x)
    n = length(x)
    τ = x[1]
    β = view(x, 2:n)
    return ((τ / 3)^2 + (n - 1) * τ + sum(b -> abs2(b * exp(-τ / 2)), β)) / -2
end
∇logp(x) = ForwardDiff.gradient(logp, x)

dim = 100
ndraws = 1_000
nruns = 20

Next, we create initial points for the Pathfinder runs.

julia> θ₀s = collect.(eachcol(rand(dim, nruns) .* 20 .- 10));

Now we run multi-path Pathfinder.

julia> @time q, ϕ = multipathfinder(logp, ∇logp, θ₀s, ndraws; ndraws_per_run=ndraws ÷ nruns);[ Info: Optimized for 794 iterations. Maximum ELBO of 83.91 reached at iteration 8.
[ Info: Optimized for 786 iterations. Maximum ELBO of 71.9 reached at iteration 18.
[ Info: Optimized for 783 iterations. Maximum ELBO of 43.29 reached at iteration 12.
[ Info: Optimized for 781 iterations. Maximum ELBO of 40.36 reached at iteration 18.
[ Info: Optimized for 798 iterations. Maximum ELBO of 88.61 reached at iteration 6.
[ Info: Optimized for 812 iterations. Maximum ELBO of 79.39 reached at iteration 10.
[ Info: Optimized for 24 iterations. Maximum ELBO of 86.48 reached at iteration 5.
[ Info: Optimized for 792 iterations. Maximum ELBO of 85.27 reached at iteration 13.
[ Info: Optimized for 803 iterations. Maximum ELBO of 86.61 reached at iteration 10.
[ Info: Optimized for 51 iterations. Maximum ELBO of 76.79 reached at iteration 14.
[ Info: Optimized for 793 iterations. Maximum ELBO of 86.45 reached at iteration 5.
[ Info: Optimized for 764 iterations. Maximum ELBO of 86.11 reached at iteration 4.
[ Info: Optimized for 797 iterations. Maximum ELBO of 75.55 reached at iteration 10.
[ Info: Optimized for 793 iterations. Maximum ELBO of 86.59 reached at iteration 9.
[ Info: Optimized for 784 iterations. Maximum ELBO of 36.6 reached at iteration 8.
[ Info: Optimized for 887 iterations. Maximum ELBO of 83.01 reached at iteration 25.
[ Info: Optimized for 797 iterations. Maximum ELBO of 89.92 reached at iteration 15.
[ Info: Optimized for 796 iterations. Maximum ELBO of 79.62 reached at iteration 12.
[ Info: Optimized for 785 iterations. Maximum ELBO of 65.07 reached at iteration 12.
[ Info: Optimized for 796 iterations. Maximum ELBO of 85.48 reached at iteration 12.
┌ Warning: Pareto k statistic exceeded 0.7. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/KhMUf/src/PSIS.jl:100
 19.799242 seconds (31.27 M allocations: 2.739 GiB, 6.25% gc time, 0.03% compilation time)

The first return value is a uniformly-weighted Distributions.MixtureModel. Each component is the result of a single Pathfinder run.

julia> typeof(q)Distributions.MixtureModel{Distributions.Multivariate, Distributions.Continuous, Distributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}, Distributions.Categorical{Float64, Vector{Float64}}}

Note that while we could draw samples from q directly, these aren't equivalent to the samples returned by multi-path Pathfinder, which uses multiple importance sampling with Pareto-smoothed importance resampling to combine the individual runs. Let's draw exact samples from the funnel and compare them with Pathfinder's samples and samples from q directly.

using Plots

τ = randn(ndraws) * 3
β₁ = @. randn() * exp(τ / 2)
τ_approx = first.(ϕ)
β₁_approx = getindex.(ϕ, 2)
ϕ2 = rand(q, ndraws)
τ_approx2 = ϕ2[1, :]
β₁_approx2 = ϕ2[2, :]

plt = scatter(β₁, τ; msw=0, ms=2, alpha=0.1, label="Exact")
scatter!(β₁_approx2, τ_approx2; msw=0, ms=2, alpha=0.2, label="multi-path Pathfinder w/o IR")
scatter!(β₁_approx, τ_approx; msw=0, ms=2, alpha=0.2, label="multi-path Pathfinder")
plot!(xlims=(-15, 15), ylims=(-15, 15), xlabel="β₁", ylabel="τ", legend=true)
plt

Here we can see that the mixture model (q) places too much probability mass on the lower part of the funnel, which is corrected by the importance resampling.