Multi-path Pathfinder

Pathfinder.multipathfinder — Function

multipathfinder(logp, ndraws; kwargs...)
multipathfinder(logp, ∇logp, ndraws; kwargs...)
multipathfinder(fun::GalacticOptim.OptimizationFunction, ndraws; kwargs...)

Run pathfinder multiple times to fit a multivariate normal mixture model.

For nruns=length(init), nruns parallel runs of pathfinder produce nruns multivariate normal approximations $q_k = q(\phi | \mu_k, \Sigma_k)$ of the posterior. These are combined to a mixture model $q$ with uniform weights.

$q$ is augmented with the component index to generate random samples, that is, elements $(k, \phi)$ are drawn from the augmented mixture model

\[\tilde{q}(\phi, k | \mu, \Sigma) = K^{-1} q(\phi | \mu_k, \Sigma_k),\]

where $k$ is a component index, and $K=$ nruns. These draws are then resampled with replacement. Discarding $k$ from the samples would reproduce draws from $q$.

If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution $p$ instead of $q$. This also prints a warning message if the importance weighted draws are unsuitable for approximating expectations with respect to $p$.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp. If not provided, logp is automatically differentiated using the backend specified in ad_backend.

fun::GalacticOptim.OptimizationFunction: an optimization function that represents -logp(x) with its gradient. It must have the necessary features (e.g. a Hessian function) for the chosen optimization algorithm. For details, see GalacticOptim.jl: OptimizationFunction.

ndraws::Int: number of approximate draws to return

Keywords

init: iterator of length nruns of initial points of length dim from which each single-path Pathfinder run will begin. length(init) must be implemented. If init is not provided, dim and nruns must be.

nruns::Int: number of runs of Pathfinder to perform. Ignored if init is provided.

ad_backend=AD.ForwardDiffBackend(): AbstractDifferentiation.jl AD backend used to differentiate logp if ∇logp is not provided.

ndraws_per_run::Int: The number of draws to take for each component before resampling. Defaults to a number such that ndraws_per_run * nruns > ndraws.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

rng::AbstractRNG=Random.GLOBAL_RNG: Pseudorandom number generator. It is recommended to use a parallelization-friendly PRNG like the default PRNG on Julia 1.7 and up.

executor::Transducers.Executor: Transducers.jl executor that determines if and how to run the single-path runs in parallel. If rng is known to be thread-safe, the default is Transducers.PreferParallel(; basesize=1) (parallel executation, defaulting to multi-threading). Otherwise, it is Transducers.SequentialEx() (no parallelization).

executor_per_run::Transducers.Executor=Transducers.SequentialEx(): Transducers.jl executor used within each run to parallelize PRNG calls. Defaults to no parallelization. See pathfinder for a description.

kwargs... : Remaining keywords are forwarded to pathfinder.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::AbstractMatrix{<:Real}: approximate draws from target distribution with size (dim, ndraws)

component_inds::Vector{Int}: Indices $k$ of components in $q$ from which each column in ϕ was drawn.

Examples

Especially for complicated target distributions, it's more useful to run multi-path Pathfinder. One difficult distribution to sample is Neal's funnel:

\[\tau \sim \mathrm{Normal}(0, 3)\\ \beta_i \sim \mathrm{Normal}(0, e^{\tau/2})\]

Such funnel geometries appear in other models and typically frustrate MCMC sampling. Multi-path Pathfinder can't sample the funnel well, but it can quickly give us draws that can help us diagnose that we have a funnel.

In this example, we draw from a 100-dimensional funnel and visualize 2 dimensions.

using ForwardDiff, Pathfinder, Random

Random.seed!(42)

function logp(x)
    n = length(x)
    τ = x[1]
    β = view(x, 2:n)
    return ((τ / 3)^2 + (n - 1) * τ + sum(b -> abs2(b * exp(-τ / 2)), β)) / -2
end
∇logp(x) = ForwardDiff.gradient(logp, x)

dim = 100
ndraws = 1_000
nruns = 20

Now we run multi-path Pathfinder.

julia> @time q, ϕ, component_ids = multipathfinder(logp, ∇logp, ndraws; init_scale=10);ERROR: ArgumentError: A positive `nruns` must be set or `init` must be provided.

The first return value is a uniformly-weighted Distributions.MixtureModel. Each component is the result of a single Pathfinder run.

julia> typeof(q)ERROR: UndefVarError: q not defined

While we could draw samples from q directly, these aren't equivalent to the samples returned by multi-path Pathfinder, which uses multiple importance sampling with Pareto-smoothed importance resampling to combine the individual runs.

Note that PSIS.jl, which smooths the importance weights, warns us that the importance weights are unsuitable for computing estimates, so we should definitely run MCMC to get better samples. Pathfinder's samples can still be useful though for initializing MCMC.

Let's compare Pathfinder's samples with samples from q directly, plotted over the exact marginal density.

using Plots

τ_approx = ϕ[1, :]
β₁_approx = ϕ[2, :]
ϕ2 = rand(q, ndraws)
τ_approx2 = ϕ2[1, :]
β₁_approx2 = ϕ2[2, :]

τ_range = -15:0.01:5
β₁_range = -5:0.01:5

plt = contour(β₁_range, τ_range, (β, τ) -> exp(logp([τ, β])))
scatter!(β₁_approx2, τ_approx2; msw=0, ms=2, alpha=0.3, color=:blue, label="Pathfinder w/o IR")
scatter!(β₁_approx, τ_approx; msw=0, ms=2, alpha=0.5, color=:red, label="Pathfinder")
plot!(xlims=extrema(β₁_range), ylims=extrema(τ_range), xlabel="β₁", ylabel="τ", legend=true)
plt

As expected from PSIS.jl's warning, neither set of samples cover the density well. But we can also see that the mixture model (q) places too many samples in regions of low density, which is corrected by the importance resampling.

We can check how much each component contributed to the returned sample.

histogram(component_ids; bins=(0:nruns) .+ 0.5, bar_width=0.8, xticks=1:nruns,
          xlabel="Component index", ylabel="Count", legend=false)