Multi-path Pathfinder

Pathfinder.multipathfinder — Function

multipathfinder(
    logp,
    ∇logp,
    θ₀s::AbstractVector{AbstractVector{<:Real}},
    ndraws::Int;
    kwargs...
)

Filter samples from a mixture of multivariate normal distributions fit using pathfinder.

For n=length(θ₀s), n parallel runs of pathfinder produce n multivariate normal approximations of the posterior. These are combined to a mixture model with uniform weights.

Draws from the components are then resampled with replacement. If importance=true, then Pareto smoothed importance resampling is used, so that the resulting draws better approximate draws from the target distribution.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀s: vector of initial points from which each optimization will begin

ndraws: number of approximate draws to return

Keywords

ndraws_per_run::Int=5: The number of draws to take for each component before resampling.

importance::Bool=true: Perform Pareto smoothed importance resampling of draws.

Returns

q::Distributions.MixtureModel: Uniformly weighted mixture of ELBO-maximizing multivariate normal distributions

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws approxiate draws from target distribution

Examples

Especially for complicated target distributions, it's more useful to run multi-path Pathfinder. One difficult distribution to sample is Neal's funnel:

\[\tau \sim \mathrm{Normal}(0, 3)\\ \beta_i \sim \mathrm{Normal}(0, e^{\tau/2})\]

Such funnel geometries appear in other models and typically frustrate MCMC sampling. Multi-path Pathfinder can't sample the funnel well, but it can quickly give us draws that can help us diagnose that we have a funnel.

In this example, we draw from a 100-dimensional funnel and visualize 2 dimensions:

julia> using ForwardDiff, Pathfinder, Plots, Random
julia> plotly()[ Info: For saving to png with the Plotly backend PlotlyBase has to be installed.
Plots.PlotlyBackend()
julia> Random.seed!(42)MersenneTwister(42)
julia> function logp(x)
           n = length(x)
           τ = x[1]
           β = view(x, 2:n)
           return ((τ / 3)^2 + (n - 1) * τ + sum(b -> abs2(b * exp(-τ / 2)), β)) / -2
       endlogp (generic function with 1 method)
julia> ∇logp(x) = ForwardDiff.gradient(logp, x);
julia> dim = 100;
julia> ndraws = 1_000;
julia> nruns = 20;
julia> θ₀s = collect.(eachcol(rand(dim, nruns) .* 20 .- 10)); # θ₀ ~ Uniform(-10, 10)
julia> @time q, ϕ = multipathfinder(logp, ∇logp, θ₀s, ndraws; ndraws_per_run=ndraws ÷ nruns);[ Info: Optimized for 795 iterations. Maximum ELBO of 83.18 reached at iteration 9.
[ Info: Optimized for 789 iterations. Maximum ELBO of 66.34 reached at iteration 19.
[ Info: Optimized for 794 iterations. Maximum ELBO of 43.62 reached at iteration 13.
[ Info: Optimized for 788 iterations. Maximum ELBO of 42.76 reached at iteration 19.
[ Info: Optimized for 791 iterations. Maximum ELBO of 88.86 reached at iteration 7.
[ Info: Optimized for 800 iterations. Maximum ELBO of 78.79 reached at iteration 11.
[ Info: Optimized for 837 iterations. Maximum ELBO of 88.53 reached at iteration 6.
[ Info: Optimized for 782 iterations. Maximum ELBO of 85.97 reached at iteration 14.
[ Info: Optimized for 796 iterations. Maximum ELBO of 85.91 reached at iteration 11.
[ Info: Optimized for 793 iterations. Maximum ELBO of 82.26 reached at iteration 11.
[ Info: Optimized for 783 iterations. Maximum ELBO of 85.9 reached at iteration 6.
[ Info: Optimized for 764 iterations. Maximum ELBO of 85.88 reached at iteration 5.
[ Info: Optimized for 40 iterations. Maximum ELBO of 83.1 reached at iteration 11.
[ Info: Optimized for 584 iterations. Maximum ELBO of 82.9 reached at iteration 10.
[ Info: Optimized for 789 iterations. Maximum ELBO of 7.05 reached at iteration 8.
[ Info: Optimized for 898 iterations. Maximum ELBO of 79.3 reached at iteration 26.
[ Info: Optimized for 794 iterations. Maximum ELBO of 89.23 reached at iteration 16.
[ Info: Optimized for 804 iterations. Maximum ELBO of 80.61 reached at iteration 13.
[ Info: Optimized for 800 iterations. Maximum ELBO of 65.81 reached at iteration 13.
[ Info: Optimized for 800 iterations. Maximum ELBO of 87.29 reached at iteration 13.
┌ Warning: Pareto k statistic exceeded 0.7. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/KhMUf/src/PSIS.jl:100
 14.924871 seconds (31.00 M allocations: 2.763 GiB, 5.49% gc time, 0.03% compilation time)
julia> typeof(q)Distributions.MixtureModel{Distributions.Multivariate, Distributions.Continuous, Distributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}, Distributions.Categorical{Float64, Vector{Float64}}}
julia> τ_approx = first.(ϕ);
julia> β₁_approx = getindex.(ϕ, 2);
julia> τ = randn(ndraws) * 3;
julia> β₁ = @. randn() * exp(τ / 2);
julia> scatter(β₁, τ; msw=0, ms=2, alpha=0.1, label="Exact");
julia> scatter!(β₁_approx, τ_approx;
               msw=0, ms=2, alpha=0.5, xlims=(-15, 15), ylims=(-15, 15),
               label="multi-path Pathfinder", xlabel="β₁", ylabel="τ", legend=true)Plot{Plots.PlotlyBackend() n=2}