Single-path Pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp[, ∇logp], θ₀::AbstractVector{<:Real}, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while maximizing logp.

From an optimization trajectory, Pathfinder constructs a sequence of (multivariate normal) approximations to the distribution specified by logp. The approximation that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between the approximation and the true distribution, is returned.

The covariance of the multivariate normal distribution is an inverse Hessian approximation constructed using at most the previous history_length steps.

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp. If not provided, logp is automatically differentiated using the backend specified in ad_backend.

θ₀: initial point of length dim from which to begin optimization

ndraws: number of approximate draws to return

Keywords

ad_backend=AD.ForwardDiffBackend(): AbstractDifferentiation.jl AD backend.

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

optimizer: Optimizer to be used for constructing trajectory. Can be any optimizer compatible with GalacticOptim, so long as it supports callbacks. Defaults to Optim.LBFGS(; m=6, linesearch=LineSearches.MoreThuente()). See the GalacticOptim.jl documentation for details.

history_length::Int=6: Size of the history used to approximate the inverse Hessian. This should only be set when optimizer is not an Optim.LBFGS.

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

kwargs... : Remaining keywords are forwarded to GalacticOptim.OptimizationProblem.

Returns

q::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

ϕ::AbstractMatrix{<:Real}: draws from multivariate normal with size (dim, ndraws)

logqϕ::Vector{<:Real}: log-density of multivariate normal at columns of ϕ

pathfinder(
    f::GalacticOptim.OptimizationFunction,
    θ₀::AbstractVector{<:Real},
    ndraws::Int;
    kwargs...,
)

Find the best multivariate normal approximation encountered while minimizing f.

f is a user-created optimization function that represents the negative log density with its gradient and must have the necessary features (e.g. a Hessian function or specified automatic differentiation type) for the chosen optimization algorithm. For details, see GalacticOptim.jl: OptimizationFunction.

See pathfinder for a description of remaining arguments.

pathfinder(prob::GalacticOptim.OptimizationProblem, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while solving prob.

prob is a user-created optimization problem that represents the negative log density with its gradient, an initial position and must have the necessary features (e.g. a Hessian function or specified automatic differentiation type) for the chosen optimization algorithm. For details, see GalacticOptim.jl: Defining OptimizationProblems.

See pathfinder for a description of remaining arguments.

Examples

For a simple example, we'll run Pathfinder on a multivariate normal distribution with a dense covariance matrix.

using LinearAlgebra, Pathfinder, Random
Random.seed!(42)

Σ = [
    2.71   0.5    0.19   0.07   1.04
    0.5    1.11  -0.08  -0.17  -0.08
    0.19  -0.08   0.26   0.07  -0.7
    0.07  -0.17   0.07   0.11  -0.21
    1.04  -0.08  -0.7   -0.21   8.65
]
μ = [-0.55, 0.49, -0.76, 0.25, 0.94]
P = inv(Symmetric(Σ))

function logp(x)
    z = x - μ
    return -dot(z, P, z) / 2;
end
∇logp(x) = P * (μ - x)

Next, we create initial points for the Pathfinder runs.

julia> θ₀ = rand(5) .* 4 .- 2  # θ₀ ~ Uniform(-2, 2);

Now we run Pathfinder.

julia> @time q, ϕ, logqϕ = pathfinder(logp, ∇logp, θ₀, 100; ndraws_elbo=100);[ Info: Optimized for 38 iterations. Maximum ELBO of 3.84 reached at iteration 34.
  7.271883 seconds (16.12 M allocations: 1.031 GiB, 6.96% gc time, 99.93% compilation time)

The first return value is a multivariate normal approximation to our target distribution. Its mean and covariance are quite close to our target distribution's.

julia> qDistributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}(
dim: 5
μ: [-0.5499999949158282, 0.4899999988178169, -0.7599999979862843, 0.2499999989895235, 0.9400000067909953]
Σ: [2.8053766005274547 0.46197980209657374 … 0.04942703113879024 1.2040248348799552; 0.46197980209657374 1.1203155589697753 … -0.1502115458530225 -0.2028405623951594; … ; 0.04942703113879024 -0.1502115458530225 … 0.12093759888809286 -0.30434396827746296; 1.2040248348799552 -0.2028405623951594 … -0.30434396827746296 9.369064525618189]
)
julia> q.μ5-element Vector{Float64}:
 -0.5499999949158282
  0.4899999988178169
 -0.7599999979862843
  0.2499999989895235
  0.9400000067909953
julia> q.Σ5×5 Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}:
 2.80538    0.46198     0.21003     0.049427    1.20402
 0.46198    1.12032    -0.0846231  -0.150212   -0.202841
 0.21003   -0.0846231   0.279982    0.0831588  -0.769305
 0.049427  -0.150212    0.0831588   0.120938   -0.304344
 1.20402   -0.202841   -0.769305   -0.304344    9.36906

ϕ is a Matrix whose columns are the requested draws from q:

julia> ϕ5×100 Matrix{Float64}:
 -0.178878    2.26743   -3.31395    …  -5.33148   -0.286022  -1.1124
 -0.0805437  -0.356872   0.210452      -0.424834   0.621246   0.718395
 -1.05981    -0.561176  -1.17655       -0.878826  -0.467025  -0.653375
  0.37856     0.425757   0.0667452     -0.198126   0.325027   0.438411
  1.14526    -0.1626    -2.57007        1.21933    1.33979   -3.02438