Single-path Pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp, ∇logp, θ₀::AbstractVector{<:Real}, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while optimizing logp.

The multivariate normal approximation returned is the one that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀: initial point of length dim from which to begin optimization

ndraws: number of approximate draws to return

Keywords

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

optimizer::Optim.AbstractOptimizer: Optimizer to be used for constructing trajectory.

Defaults to Optim.LBFGS(; m=6, linesearch=LineSearches.MoreThuente{Float64} f_tol: Float64 0.0001 gtol: Float64 0.9 x_tol: Float64 1.0e-8 alphamin: Float64 1.0e-16 alphamax: Float64 65536.0 maxfev: Int64 100 ).

history_length::Int=6: Size of the history used to approximate the

inverse Hessian. This should only be set when optimizer is not an Optim.LBFGS.

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

kwargs... : Remaining keywords are forwarded to Optim.Options.

Returns

q::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

ϕ::AbstractMatrix{<:Real}: draws from multivariate normal with size (dim, ndraws)

logqϕ::Vector{<:Real}: log-density of multivariate normal at columns of ϕ

Examples

For a simple example, we'll run Pathfinder on a multivariate normal distribution with a dense covariance matrix.

using LinearAlgebra, Pathfinder, Random
Random.seed!(42)

Σ = [
    2.71   0.5    0.19   0.07   1.04
    0.5    1.11  -0.08  -0.17  -0.08
    0.19  -0.08   0.26   0.07  -0.7
    0.07  -0.17   0.07   0.11  -0.21
    1.04  -0.08  -0.7   -0.21   8.65
]
μ = [-0.55, 0.49, -0.76, 0.25, 0.94]
P = inv(Symmetric(Σ))

function logp(x)
    z = x - μ
    return -dot(z, P, z) / 2;
end
∇logp(x) = P * (μ - x)

Next, we create initial points for the Pathfinder runs.

julia> θ₀ = rand(5) .* 4 .- 2  # θ₀ ~ Uniform(-2, 2);

Now we run Pathfinder.

julia> @time q, ϕ, logqϕ = pathfinder(logp, ∇logp, θ₀, 100; ndraws_elbo=100);[ Info: Optimized for 19 iterations. Maximum ELBO of 3.84 reached at iteration 14.
  0.544206 seconds (953.02 k allocations: 56.370 MiB, 99.34% compilation time)

The first return value is a multivariate normal approximation to our target distribution. Its mean and covariance are quite close to our target distribution's.

julia> qDistributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}(
dim: 5
μ: [-0.5499998961951266, 0.4900000565078103, -0.759999946178205, 0.250000017692644, 0.9399995211722131]
Σ: [2.6865223104978826 0.4913100508451975 … 0.07492542674541437 0.948632317938157; 0.4913100508451975 1.1070966060677372 … -0.16749367736822252 -0.1291001914369837; … ; 0.07492542674541437 -0.16749367736822252 … 0.11045777614630284 -0.22418213942978152; 0.948632317938157 -0.1291001914369837 … -0.22418213942978152 9.040353112322498]
)
julia> q.μ5-element Vector{Float64}:
 -0.5499998961951266
  0.4900000565078103
 -0.759999946178205
  0.250000017692644
  0.9399995211722131
julia> q.Σ5×5 Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}:
 2.68652     0.49131     0.197092    0.0749254   0.948632
 0.49131     1.1071     -0.0759006  -0.167494   -0.1291
 0.197092   -0.0759006   0.264802    0.07173    -0.744372
 0.0749254  -0.167494    0.07173     0.110458   -0.224182
 0.948632   -0.1291     -0.744372   -0.224182    9.04035

ϕ is a Matrix whose columns are the requested draws from q:

julia> ϕ5×100 Matrix{Float64}:
  0.322592   1.23638   -1.28841    …   1.91746    3.14378    -0.101451
 -0.166008   0.28346   -2.42926        0.161913   0.73329     0.307336
 -0.566246  -0.718596  -0.0557306     -1.05968   -0.0686237  -0.907436
  0.783459   0.36381    0.816146       0.845115   0.71726     0.454126
 -1.98377    2.77225   -0.636019       2.19535    2.44898     1.81705