Single-path Pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp, ∇logp, θ₀::AbstractVector{<:Real}, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while optimizing logp.

The multivariate normal approximation returned is the one that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀: initial point from which to begin optimization

ndraws: number of approximate draws to return

Keywords

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

optimizer::Optim.AbstractOptimizer: Optimizer to be used for constructing trajectory.

Defaults to Optim.LBFGS(; m=5, linesearch=LineSearches.MoreThuente()).

history_length::Int=5: Size of the history used to approximate the inverse Hessian.

This should only be set when optimizer is not an Optim.LBFGS.

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

kwargs... : Remaining keywords are forwarded to Optim.Options.

Returns

q::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws draws from multivariate normal

logqϕ::Vector{<:Real}: log-density of multivariate normal at ϕ values

Examples

For a simple example, we'll run Pathfinder on a multivariate normal distribution with a dense covariance matrix.

using LinearAlgebra, Pathfinder, Random
Random.seed!(42)

Σ = [
    2.71   0.5    0.19   0.07   1.04
    0.5    1.11  -0.08  -0.17  -0.08
    0.19  -0.08   0.26   0.07  -0.7
    0.07  -0.17   0.07   0.11  -0.21
    1.04  -0.08  -0.7   -0.21   8.65
]
μ = [-0.55, 0.49, -0.76, 0.25, 0.94]
P = inv(Symmetric(Σ))

function logp(x)
    z = x - μ
    return -dot(z, P, z) / 2;
end
∇logp(x) = P * (μ - x)

Next, we create initial points for the Pathfinder runs.

julia> θ₀ = rand(5) .* 4 .- 2  # θ₀ ~ Uniform(-2, 2);

Now we run Pathfinder.

julia> @time q, ϕ, logqϕ = pathfinder(logp, ∇logp, θ₀, 100; ndraws_elbo=100);[ Info: Optimized for 26 iterations. Maximum ELBO of 3.82 reached at iteration 20.
  0.691588 seconds (985.18 k allocations: 58.556 MiB, 3.10% gc time, 99.20% compilation time)

The first return value is a multivariate normal approximation to our target distribution. Its mean and covariance are quite close to our target distribution's.

julia> qDistributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}(
dim: 5
μ: [-0.5500041404453908, 0.4900000795117144, -0.7600005214906455, 0.2499997929610976, 0.9400004528346747]
Σ: [4.120830304187059 0.45445740302283083 … 0.14374179747283614 0.88341352511756; 0.45445740302283083 1.1459025337468847 … -0.1831362012682332 -0.0981451233813781; … ; 0.14374179747283614 -0.1831362012682332 … 0.11645116897363425 -0.21530932615911783; 0.88341352511756 -0.0981451233813781 … -0.21530932615911783 8.65796167738062]
)
julia> q.μ5-element Vector{Float64}:
 -0.5500041404453908
  0.4900000795117144
 -0.7600005214906455
  0.2499997929610976
  0.9400004528346747
julia> q.Σ5×5 Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}:
 4.12083    0.454457    0.368989    0.143742    0.883414
 0.454457   1.1459     -0.0872493  -0.183136   -0.0981451
 0.368989  -0.0872493   0.282741    0.0799654  -0.718013
 0.143742  -0.183136    0.0799654   0.116451   -0.215309
 0.883414  -0.0981451  -0.718013   -0.215309    8.65796