Single-path Pathfinder

Pathfinder.pathfinder — Function

pathfinder(logp, ∇logp, θ₀::AbstractVector{<:Real}, ndraws::Int; kwargs...)

Find the best multivariate normal approximation encountered while optimizing logp.

The multivariate normal approximation returned is the one that maximizes the evidence lower bound (ELBO), or equivalently, minimizes the KL divergence between

Arguments

logp: a callable that computes the log-density of the target distribution.

∇logp: a callable that computes the gradient of logp.

θ₀: initial point from which to begin optimization

ndraws: number of approximate draws to return

Keywords

rng::Random.AbstractRNG: The random number generator to be used for drawing samples

optimizer::Optim.AbstractOptimizer: Optimizer to be used for constructing trajectory.

Defaults to Optim.LBFGS(; m=5, linesearch=LineSearches.MoreThuente()).

history_length::Int=5: Size of the history used to approximate the inverse Hessian.

This should only be set when optimizer is not an Optim.LBFGS.

ndraws_elbo::Int=5: Number of draws used to estimate the ELBO

kwargs... : Remaining keywords are forwarded to Optim.Options.

Returns

q::Distributions.MvNormal: ELBO-maximizing multivariate normal distribution

ϕ::Vector{<:AbstractVector{<:Real}}: ndraws draws from multivariate normal

logqϕ::Vector{<:Real}: log-density of multivariate normal at ϕ values

Examples

For a simple example, we'll run Pathfinder on a multivariate normal distribution with a dense covariance matrix.

julia> using LinearAlgebra, Pathfinder, Random
julia> Random.seed!(42)MersenneTwister(42)
julia> Σ = [
           2.71   0.5    0.19   0.07   1.04
           0.5    1.11  -0.08  -0.17  -0.08
           0.19  -0.08   0.26   0.07  -0.7
           0.07  -0.17   0.07   0.11  -0.21
           1.04  -0.08  -0.7   -0.21   8.65
       ];
julia> μ = [-0.55, 0.49, -0.76, 0.25, 0.94];
julia> P = inv(Symmetric(Σ));
julia> function logp(x)
           z = x - μ
           return -dot(z, P, z) / 2;
       end;
julia> ∇logp(x) = P * (μ - x);
julia> θ₀ = rand(5) .* 4 .- 2  # θ₀ ~ Uniform(-2, 2)5-element Vector{Float64}:
  0.13273206417544525
 -0.18388345765143033
 -1.9292526931401426
 -1.3082678842521949
  1.8357035053189392[ Info: Optimized for 26 iterations. Maximum ELBO of 3.81 reached at iteration 11.
julia> @time q, ϕ, logqϕ = pathfinder(logp, ∇logp, θ₀, 100; ndraws_elbo=100);[ Info: Optimized for 26 iterations. Maximum ELBO of 3.87 reached at iteration 22.
  0.013821 seconds (18.29 k allocations: 1.428 MiB, 85.53% compilation time)
julia> qDistributions.MvNormal{Float64, Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}, Vector{Float64}}(
dim: 5
μ: [-0.5499999330965111, 0.4900000697341321, -0.7599999716965999, 0.2500000002232026, 0.9400000946307552]
Σ: [2.6098302533853195 0.5535479608145211 … 0.03924833492672704 0.8671193716850419; 0.5535479608145211 0.892816420125516 … -0.11153732092075373 -0.0034619220099992845; … ; 0.03924833492672704 -0.11153732092075373 … 0.09338900391688362 -0.2799437727305112; 0.8671193716850419 -0.0034619220099992845 … -0.2799437727305112 8.540019840274025]
)
julia> q.μ5-element Vector{Float64}:
 -0.5499999330965111
  0.4900000697341321
 -0.7599999716965999
  0.2500000002232026
  0.9400000946307552
julia> q.Σ5×5 Pathfinder.WoodburyPDMat{Float64, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, Matrix{Float64}, Matrix{Float64}, LinearAlgebra.Diagonal{Float64, Vector{Float64}}, LinearAlgebra.QRCompactWYQ{Float64, Matrix{Float64}}, LinearAlgebra.UpperTriangular{Float64, Matrix{Float64}}}:
 2.60983     0.553548     0.132744    0.0392483   0.867119
 0.553548    0.892816    -0.0650377  -0.111537   -0.00346192
 0.132744   -0.0650377    0.301141    0.0430584  -0.680652
 0.0392483  -0.111537     0.0430584   0.093389   -0.279944
 0.867119   -0.00346192  -0.680652   -0.279944    8.54002